+++ zee-social-buttons.js
@@ -1,10 +1,10 @@
 /*
  * Zee Social Buttons
- * Version: 2.5
+ * Version: 2.8
  * Develop By ClubCoding
  * Download/Clone: https://github.com/krishnaTORQUE/Zee-Social-Buttons
  */
-window.addEventListener('load', function () {
+function zee_social_buttons() {
 
     var zee_social_button_img = null;
 
@@ -23,22 +23,23 @@ window.addEventListener('load', function () {
      */
 
     var zee_social_buttons_elms = {
-        facebook: '<div class="zee_social_sbutton zee_sbtn_facebook" title="Facebook"></div>',
-        twitter: '<div class="zee_social_sbutton zee_sbtn_twitter" title="Twitter"></div>',
-        googleplus: '<div class="zee_social_sbutton zee_sbtn_googleplus" title="Google+"></div>',
-        linkedin: '<div class="zee_social_sbutton zee_sbtn_linkedin" title="LinkedIn"></div>',
-        pinterest: '<div class="zee_social_sbutton zee_sbtn_pinterest" title="Pinterest"></div>',
-        stumbleupon: '<div class="zee_social_sbutton zee_sbtn_stumbleupon" title="Stumbleupon"></div>',
-        buffer: '<div class="zee_social_sbutton zee_sbtn_buffer" title="Buffer"></div>',
-        reddit: '<div class="zee_social_sbutton zee_sbtn_reddit" title="Reddit"></div>',
-        pocket: '<div class="zee_social_sbutton zee_sbtn_pocket" title="Pocket"></div>',
-        email: '<div class="zee_social_sbutton zee_sbtn_email" title="Email"></div>',
+        facebook: '<div class="zee_social_sbutton zee_sbtn_facebook" title="Facebook">&nbsp;</div>',
+        twitter: '<div class="zee_social_sbutton zee_sbtn_twitter" title="Twitter">&nbsp;</div>',
+        googleplus: '<div class="zee_social_sbutton zee_sbtn_googleplus" title="Google+">&nbsp;</div>',
+        linkedin: '<div class="zee_social_sbutton zee_sbtn_linkedin" title="LinkedIn">&nbsp;</div>',
+        pinterest: '<div class="zee_social_sbutton zee_sbtn_pinterest" title="Pinterest">&nbsp;</div>',
+        stumbleupon: '<div class="zee_social_sbutton zee_sbtn_stumbleupon" title="Stumbleupon">&nbsp;</div>',
+        buffer: '<div class="zee_social_sbutton zee_sbtn_buffer" title="Buffer">&nbsp;</div>',
+        reddit: '<div class="zee_social_sbutton zee_sbtn_reddit" title="Reddit">&nbsp;</div>',
+        pocket: '<div class="zee_social_sbutton zee_sbtn_pocket" title="Pocket">&nbsp;</div>',
+        whatsapp: '<div class="zee_social_sbutton zee_sbtn_whatsapp" title="WhatsApp">&nbsp;</div>',
+        email: '<div class="zee_social_sbutton zee_sbtn_email" title="Email">&nbsp;</div>'
     };
 
     var zee_social_buttons = document.getElementsByClassName('zee-social-buttons');
     for (var i = 0; i < zee_social_buttons.length; i++) {
 
-        var total_social_buttons, zee_social_buttons_content = '';
+        var total_social_buttons, zee_social_buttons_content = '<div class="zee_social_sbutton zee_sbtn_count" title="Total Share">. . .&emsp;</div>';
 
         /*
          * Loading giving buttons
@@ -79,7 +80,7 @@ window.addEventListener('load', function () {
          * Calculaion Width
          */
 
-        total_social_buttons = (100 / total_social_buttons);
+        total_social_buttons = 100 / (parseInt(total_social_buttons) + 1);
 
         /*
          * Put Contents in Body
@@ -154,6 +155,10 @@ window.addEventListener('load', function () {
                         social_buttons_link = 'https://getpocket.com/save?url=' + zee_social_buttons_share_details['url_en'] + '&title=' + zee_social_buttons_share_details['title'];
                         break;
 
+                    case 'WhatsApp':
+                        social_buttons_link = '"whatsapp://send?data-link=' + zee_social_buttons_share_details['url_en'] + '&data-text=' + zee_social_buttons_share_details['title_en'];
+                        break;
+
                     case 'Email':
                         social_buttons_link = 'mailto:?subject=' + zee_social_buttons_share_details['title_en'] + '&body=Check this out: ' + zee_social_buttons_share_details['url_en'];
                         break;
@@ -168,10 +173,43 @@ window.addEventListener('load', function () {
         }
 
         /*
-         * Finally Showing Social Buttons
+         * Getting Total Count
+         */
+
+        var theCount = [];
+        var getCountLink = 'http://public.newsharecounts.com/count.json?url=' + zee_social_buttons_share_details['url_en'] + '&providers=facebook,twitter,google,linkedin,pinterest,stumbleupon,buffer,pocket&format=jsonp';
+        var xhttp = new XMLHttpRequest();
+        xhttp.onreadystatechange = function () {
+            if (this.readyState === 4 && this.status === 200) {
+                theCount = JSON.parse(this.responseText.trim());
+            }
+        };
+        xhttp.open("GET", getCountLink, true);
+        xhttp.send();
+
+        /*
+         * Showing Total Count
+         */
+
+        setTimeout(function () {
+            if (theCount['count'] !== 'undefined') {
+                var zee_sbtn_count = document.getElementsByClassName('zee_sbtn_count');
+                for (var l = 0; l < zee_sbtn_count.length; l++) {
+                    zee_sbtn_count[l].innerHTML = theCount['count'] + '&emsp;';
+                }
+            } else {
+                zee_sbtn_count[l].innerHTML = '0&emsp;';
+            }
+        }, 1500);
+
+        /*
+         * Final Render Social Buttons
          */
 
         zee_social_buttons[i].style.visibility = 'visible';
 
     }
-}, false);
+}
+
+window.addEventListener('load', zee_social_buttons, false);
+window.addEventListener('resize', zee_social_buttons, false);